# docs/swagger.yaml
basePath: /
definitions:
  genprotos.Education:
    properties:
      degree:
        type: string
      end_date:
        type: string
      field_of_study:
        type: string
      id:
        type: string
      institution:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  genprotos.Experience:
    properties:
      company:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
  genprotos.GetAllProjects:
    properties:
      projects:
        items:
          $ref: '#/definitions/genprotos.Project'
        type: array
    type: object
  genprotos.GetAllSkills:
    properties:
      skills:
        items:
          $ref: '#/definitions/genprotos.Skill'
        type: array
    type: object
  genprotos.Project:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
      url:
        type: string
      user_id:
        type: string
    type: object
  genprotos.Skill:
    properties:
      id:
        type: string
      level:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
  genprotos.UserPortfolio:
    properties:
      educations:
        items:
          $ref: '#/definitions/genprotos.Education'
        type: array
      experiences:
        items:
          $ref: '#/definitions/genprotos.Experience'
        type: array
      projects:
        items:
          $ref: '#/definitions/genprotos.Project'
        type: array
      skills:
        items:
          $ref: '#/definitions/genprotos.Skill'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: Portfolio
  version: "1.0"
paths:
...
  /portfolio/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve user portfolio by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get User Portfolio Successful
          schema:
            $ref: '#/definitions/genprotos.UserPortfolio'
        "400":
          description: User ID is required
          schema:
            type: string
        "500":
          description: Error while retrieving portfolio
          schema:
            type: string
      summary: Get User Portfolio
      tags:
      - Portfolio
  /project/byuser:
    get:
      consumes:
      - application/json
      description: Retrieve projects by user ID
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get Projects by User ID Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllProjects'
        "400":
          description: User ID is required
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error while retrieving projects
          schema:
            type: string
      summary: Get Projects by User ID
      tags:
      - Project
  /project/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Project'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Project
      tags:
      - Project
  /project/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Project
      tags:
      - Project
  /project/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      parameters:
      - in: query
        name: description
        type: string
      - in: query
        name: id
        type: string
      - in: query
        name: title
        type: string
      - in: query
        name: url
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllProjects'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Project
      tags:
      - Project
  /project/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Project'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Project
      tags:
      - Project
  /project/search:
    get:
      consumes:
      - application/json
      description: Search projects by name or description
      parameters:
      - description: Search Query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllProjects'
        "400":
          description: Search query is required
          schema:
            type: string
        "500":
          description: Error while searching projects
          schema:
            type: string
      summary: Search Projects
      tags:
      - Project
  /project/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Project'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Project
      tags:
      - Project
  /skill/byuser/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve skills by user ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Get Skills by User ID Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllSkills'
        "400":
          description: User ID is required
          schema:
            type: string
        "404":
          description: Skills not found
          schema:
            type: string
        "500":
          description: Error while retrieving skills
          schema:
            type: string
      summary: Get Skills by User ID
      tags:
      - Skill
  /skill/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.Skill'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create Porfolio
      tags:
      - Skill
  /skill/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete Skill
      tags:
      - Skill
  /skill/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      parameters:
      - in: query
        name: id
        type: string
      - in: query
        name: level
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllSkills'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll Skill
      tags:
      - Skill
  /skill/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.Skill'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById Skill
      tags:
      - Skill
  /skill/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.Skill'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update Skill
      tags:
      - Skill
swagger: "2.0"